requirements.txt
Django==4.1.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.5
drf-yasg==1.21.4
python-decouple==3.7
gunicorn==20.1.0
django-simple-history==3.2.0
